requirements.txt
django
pillow
python-dotenv
requests
djangorestframework
markdown
django-filter